Renamed util.binary to util.string.
